+++ package/Editor/Utils.cs
@@ -230,4 +230,30 @@ namespace Unity.QuickSearch
         public static readonly int Minor;
         public static readonly int Build;
     }
+
+    internal struct BlinkCursorScope : IDisposable
+    {
+        private bool changed;
+        private Color oldCursorColor;
+
+        public BlinkCursorScope(bool blink, Color blinkColor)
+        {
+            changed = false;
+            oldCursorColor = Color.white;
+            if (blink)
+            {
+                oldCursorColor = GUI.skin.settings.cursorColor;
+                GUI.skin.settings.cursorColor = blinkColor;
+                changed = true;
+            }
+        }
+
+        public void Dispose()
+        {
+            if (changed)
+            {
+                GUI.skin.settings.cursorColor = oldCursorColor;
+            }
+        }
+    }
 }